refactor: include parsed operation title to notification
