+++ src/renderer/domains/network/multisig-operation/service.ts
@@ -8,14 +8,10 @@ import {
   type CallHash,
   type Chain,
   ChainOptions,
-  type CreateMultisigEventParams,
-  type CreateMultisigOperationParams,
   CryptoType,
   type FlexibleMultisigAccount,
   type MultisigAccount,
   type NoID,
-  type NotificationStatus,
-  NotificationType,
   type Serializable,
 } from '@/shared/core';
 import { isEqual, merge, nonNullable, nullable, validateCallData } from '@/shared/lib/utils';
@@ -24,7 +20,7 @@ import { Paths } from '@/shared/routes';
 import { transactionService } from '../transaction/service';
 
 import { DEFAULT_BLOCK_HASH, MULTISIG_EXTRINSIC_CALL_INDEX, WRAP_EXTRINSIC_CALL_INDEX } from './constants';
-import { type MultisigEvent, type MultisigOperation } from './types';
+import { type MultisigOperation } from './types';
 
 /**
  * Public keys of signers' wallets are compared byte-for-byte and sorted
@@ -190,118 +186,6 @@ function generateMultisigOperationRelativeLink(params: MultisigOperationDeepLink
   return `${Paths.OPERATIONS}?${searchParams.toString()}`;
 }
 
-const getNotificationStatus = (operationStatus: 'pending' | 'executed' | 'cancelled' | 'error'): NotificationStatus => {
-  switch (operationStatus) {
-    case 'pending':
-      return 'info';
-    case 'executed':
-      return 'success';
-    case 'cancelled':
-    case 'error':
-      return 'error';
-  }
-};
-
-const getNotificationTitle = (operationStatus: 'pending' | 'executed' | 'cancelled' | 'error'): string => {
-  switch (operationStatus) {
-    case 'pending':
-      return 'Multisig operation created';
-    case 'executed':
-      return 'Multisig operation executed';
-    case 'cancelled':
-      return 'Multisig operation rejected';
-    case 'error':
-      return 'Multisig operation error';
-  }
-};
-
-const createOperationNotification = (
-  operation: MultisigOperation,
-  walletName?: string,
-): CreateMultisigOperationParams => {
-  const description = walletName ? `by ${walletName}` : undefined;
-
-  const relativeLink = multisigOperationService.generateMultisigOperationRelativeLink({
-    chainId: operation.chainId,
-    callHash: operation.callHash,
-    accountId: operation.accountId,
-    blockCreated: operation.blockCreated,
-    indexCreated: operation.indexCreated,
-  });
-
-  return {
-    key: `${NotificationType.MULTISIG_OPERATION}-${multisigOperationService.getOperationId(operation.chainId, operation.callHash, operation.accountId, operation.blockCreated, operation.indexCreated)}-${operation.status}`,
-    type: NotificationType.MULTISIG_OPERATION,
-    status: getNotificationStatus(operation.status),
-    issuer: operation.accountId,
-    title: getNotificationTitle(operation.status),
-    description,
-    multisigAccountId: operation.accountId,
-    callHash: operation.callHash,
-    callTimepoint: {
-      height: operation.blockCreated,
-      index: operation.indexCreated,
-    },
-    chainId: operation.chainId,
-    link: {
-      title: 'notifications.details.viewOperation',
-      path: relativeLink,
-    },
-    batch: {
-      title: 'notifications.toast.batch.multisigOperationsUpdated',
-      link: {
-        title: 'notifications.toast.viewOperations',
-        path: Paths.OPERATIONS,
-      },
-    },
-  };
-};
-
-const createEventNotification = (
-  operation: MultisigOperation,
-  event: MultisigEvent,
-  signerName?: string,
-): CreateMultisigEventParams => {
-  const relativeLink = multisigOperationService.generateMultisigOperationRelativeLink({
-    chainId: operation.chainId,
-    callHash: operation.callHash,
-    accountId: operation.accountId,
-    blockCreated: operation.blockCreated,
-    indexCreated: operation.indexCreated,
-  });
-
-  const eventId = `${multisigOperationService.getOperationId(operation.chainId, operation.callHash, operation.accountId, operation.blockCreated, operation.indexCreated)}-${event.id}`;
-
-  return {
-    key: `${NotificationType.MULTISIG_EVENT}-${eventId}`,
-    type: NotificationType.MULTISIG_EVENT,
-    status: event.status === 'approve' ? 'success' : 'error',
-    issuer: operation.accountId,
-    title: event.status === 'approve' ? 'Multisig operation signed' : 'Multisig operation rejected',
-    description: signerName ? `by ${signerName}` : undefined,
-    multisigAccountId: operation.accountId,
-    callHash: operation.callHash,
-    callTimepoint: {
-      height: operation.blockCreated,
-      index: operation.indexCreated,
-    },
-    chainId: operation.chainId,
-    signerAccountId: event.accountId,
-    eventStatus: event.status,
-    link: {
-      title: 'notifications.details.viewOperation',
-      path: relativeLink,
-    },
-    batch: {
-      title: 'notifications.toast.batch.multisigEventsUpdated',
-      link: {
-        title: 'notifications.toast.viewOperations',
-        path: Paths.OPERATIONS,
-      },
-    },
-  };
-};
-
 export const multisigOperationService = {
   getOperationId,
   getEventId,
@@ -315,7 +199,4 @@ export const multisigOperationService = {
 
   generateMultisigOperationDeepLink,
   generateMultisigOperationRelativeLink,
-
-  createOperationNotification,
-  createEventNotification,
 };
